pyproject.toml
[project]
name = "dataset-analyzer"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "pandas", # Keep pandas for now, as main.py still uses it for df.to_string()
    "polars",
    "langchain-google-genai",
    "langchain",
    "langchain-core",
    "python-dotenv",
    "openpyxl",
    "openai",
    "groq",
    "transformers",
    "torch",
    "cerebras-cloud-sdk>=1.49.0",
    "chromadb",
    "onnxruntime==1.16.3",
    "scikit-learn>=1.6.1",
]

[tool.setuptools.packages.find]
include = ["dataset_analyzer", "tools"]

